+++ src/Util.ts
@@ -2,7 +2,7 @@
 
 import * as path from 'path'
 import * as fs from 'fs'
-import { Uri } from 'vscode';
+import { RelativePattern, Uri, window, workspace, WorkspaceFolder } from 'vscode';
 import { DocumentUri } from 'vscode-languageclient';
 
 export function ensureDirectoryExistence(filePath) {
@@ -11,9 +11,22 @@ export function ensureDirectoryExistence(filePath) {
         return true;
     }
 
-    fs.mkdirSync(dirname);
+    fs.mkdirSync(dirname,{recursive : true});
     
     return fs.existsSync(dirname);
+}
+
+export function getDefaultWorkspaceFolder(): WorkspaceFolder | undefined {
+    if (workspace.workspaceFolders === undefined) {
+        return undefined;
+    }
+    if (workspace.workspaceFolders.length === 1) {
+        return workspace.workspaceFolders[0];
+    }
+    if (window.activeTextEditor) {
+        return workspace.getWorkspaceFolder(window.activeTextEditor.document.uri);
+    }
+    return undefined;
 }
 
 export function getJarsFromFolder(resourcesPath: string): string[]{
@@ -53,19 +66,18 @@ export function isDir(path: fs.PathLike): boolean {
     return fs.lstatSync(path).isDirectory();
 }
 
-export function createLibDirectory(rootPath: Uri): Promise<DocumentUri>{
-    return new Promise( async (resolve, reject) => { // TODO any reason for the "async" keyword here?
-        let fullUri = Uri.joinPath(rootPath, "lib");
+export function createDirectory(fullUri: Uri): Promise<void>{
+    return new Promise( (resolve, reject) => { 
         ensureDirectoryExistence(fullUri.fsPath);
         fs.access(fullUri.fsPath, fs.constants.F_OK | fs.constants.R_OK, (accessErr) => {
             if(!accessErr)
-                return resolve(fullUri.fsPath);
+                return resolve();
             if (accessErr.code === 'ENOENT'){
                 fs.mkdir(fullUri.fsPath, dirErr => {
                     if (dirErr){
                         return reject(dirErr);
                     }
-                    return resolve(fullUri.fsPath); // TODO this was .toString() before, was there a reason for that?
+                    return resolve();
                 });     
             }
             else
@@ -132,4 +144,16 @@ export function findJavaExecutable(binname: string) {
 
     // Else return the binary name directly (this will likely always fail downstream) 
     return null;
+}
+
+export async function guessDialect(wsFolder: WorkspaceFolder){
+    const dialects = {"vdmsl" : "SL", "vdmpp" : "PP", "vdmrt" : "RT"}
+    let dialect = undefined;
+    for (var dp in dialects){
+        let pattern = new RelativePattern(wsFolder.uri.path, "*." + dp);
+        let res = await workspace.findFiles(pattern,null,1)
+        if(res.length == 1) dialect = dp;
+    }
+    
+    return dialect
 }