+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 1.2.0
+- Improved outline view
+- Fixes to MATH lib
+- Added import example projects
+- Added CSV lib support
+- Added more debug configuration snippets
+
 ### 1.1.3
 - Added keywords to syntax highlighting
 - Fixes to VDMJ